Feature/diary

+++ src/main/java/com/arom/yeojung/service/UserDiaryService.java
@@ -1,9 +1,6 @@
 package com.arom.yeojung.service;
 
-import com.arom.yeojung.object.BaseTimeEntity;
-import com.arom.yeojung.object.Diary;
-import com.arom.yeojung.object.User;
-import com.arom.yeojung.object.UserDiary;
+import com.arom.yeojung.object.*;
 import com.arom.yeojung.object.dto.UserDiaryDto;
 import com.arom.yeojung.repository.DiaryRepository;
 import com.arom.yeojung.repository.UserDiaryRepository;
@@ -28,14 +25,14 @@ public class UserDiaryService extends BaseTimeEntity {
 
     //특정 사용자의 모든 다이어리 조회
     public List<UserDiaryDto> getUserDiaryByUserId(Long userId) {
-        List<UserDiary> userDiaries = userDiaryRepository.findAllByUserId(userId)
+        List<UserDiary> userDiaries = userDiaryRepository.findAllByUser_UserId(userId)
                 .orElseThrow(() -> new CustomException(ErrorCode.USERDIAYR_NOT_FOUND));
         return userDiaries.stream().map(UserDiary::EntityToDto).collect(Collectors.toList());
     }
 
     //특정 다이어리에 속한 사용자 목록 조회
     public List<UserDiaryDto> getUserDiaryByDiaryId(Long diaryId) {
-        List<UserDiary> userDiaries = userDiaryRepository.findAllByDiaryId(diaryId)
+        List<UserDiary> userDiaries = userDiaryRepository.findAllByDiary_DiaryId(diaryId)
                 .orElseThrow(() -> new CustomException(ErrorCode.USERDIAYR_NOT_FOUND));
         return userDiaries.stream().map(UserDiary::EntityToDto).collect(Collectors.toList());
     }
@@ -56,9 +53,13 @@ public class UserDiaryService extends BaseTimeEntity {
     }
 
     //사용자가 다이어리에서 나가기
-    public String removeUerFromDiary(Long userId, Long diaryId) {
-        UserDiary userDiary = userDiaryRepository.findByUserIdAndDiaryId(userId, diaryId)
+    public String removeUerFromDiary(Long userId, Long diaryId, User currentUser) {
+        UserDiary userDiary = userDiaryRepository.findByUser_UserIdAndDiary_DiaryId(userId, diaryId)
                 .orElseThrow(() -> new CustomException(ErrorCode.USERDIAYR_NOT_FOUND));
+
+        if(!userRepository.findByUserId(userId).equals(currentUser)) {
+            throw new CustomException(ErrorCode.ACCESS_DENIED);
+        }
 
         userDiaryRepository.delete(userDiary);
         return "remove User From Diary Success";
